Deleted extensible-ui files that had been moved.
- Also merged thinkyhead's modifications to the originals.
