--- Marlin/src/extensible-ui/lib/dummy.cpp
@@ -1,44 +1,0 @@
-/*************
- * dummy.cpp *
- *************/
-
-/****************************************************************************
- *   Written By Marcio Teixeira 2018 - Aleph Objects, Inc.                  *
- *                                                                          *
- *   This program is free software: you can redistribute it and/or modify   *
- *   it under the terms of the GNU General Public License as published by   *
- *   the Free Software Foundation, either version 3 of the License, or      *
- *   (at your option) any later version.                                    *
- *                                                                          *
- *   This program is distributed in the hope that it will be useful,        *
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of         *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the          *
- *   GNU General Public License for more details.                           *
- *                                                                          *
- *   To view a copy of the GNU General Public License, go to the following  *
- *   location: <http://www.gnu.org/licenses/>.                              *
- ****************************************************************************/
-
-#include "../../inc/MarlinConfigPre.h"
-
-#if ENABLED(EXTENSIBLE_UI)
-
-#include "../ui_api.h"
-
-// To implement a new UI, complete the functions below and
-// read or update Marlin's state using the methods in the
-// Extensible_UI_API methods in "../ui_api.h"
-
-namespace Extensible_UI_API {
-  void onStartup() {}
-  void onIdle() {}
-  void onUpdate() {}
-  void onPrinterKilled(const char* lcd_msg) {}
-  void onMediaInserted() {}
-  void onMediaRemoved() {}
-  void onPlayTone(const uint16_t frequency, const uint16_t duration) {}
-  void onStatusChanged(const char* lcd_msg) {}
-  void onStatusChanged(progmem_str lcd_msg) {}
-}
-
-#endif // EXTENSIBLE_UI
